+++ fabikeeper/routes/base_route.py
@@ -1,0 +1,10 @@
+from flask import Blueprint, render_template
+
+NAME = 'BASE'
+
+bp = Blueprint(NAME, __name__)
+
+
+@bp.route('/')
+def index():
+    return render_template('index.html')
