+++ fabikeeper/templates/auth/login.html
@@ -1,4 +1,4 @@
-<form method="POST" action="{{ url_for('login') }}">
+<form method="POST" action="{{ url_for('auth.login') }}">
     {{ form.csrf_token }}
     {{ form.user_id.label }} {{ form.user_id(minlength=4, maxlength=20) }}
     {{ form.password.label }} {{ form.password(minlength=4, maxlength=20) }}
