+++ fabikeeper/templates/auth/register.html
@@ -1,4 +1,4 @@
-<form method="POST" action="{{ url_for('register') }}">
+<form method="POST" action="{{ url_for('auth.register') }}">
     {{ form.csrf_token }}
     {{ form.user_id.label }} {{ form.user_id(minlength=4, maxlength=20) }}
     {{ form.user_name.label }} {{ form.user_name(minlength=4, maxlength=20) }}
